Update plugin documentation and tests
- Updated PLUGIN_SYSTEM_PROGRESS.md with next steps
- Fixed test script for Azure plugin
- Modified mock Azure provider for better testing
- Added integration test entry in roadmap
